fleet-formationのfleetのアダプター構成をリファクタリング
