+++ fleet-formation/pkg/fleet/adapters/rabbitmq/fleet_id_removed_event_handler_test.go
@@ -1,0 +1,40 @@
+package rabbitmq
+
+import (
+	"fleet-formation/pkg/fleet/app"
+	"testing"
+
+	"github.com/Tomofiles/skysign_cloud_v2/skysign-proto/pkg/skysign_proto"
+
+	"github.com/golang/protobuf/proto"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+)
+
+func TestHandleFleetIDRemovedEvent(t *testing.T) {
+	a := assert.New(t)
+
+	service := manageFleetServiceMock{}
+
+	service.On("DeleteFleet", mock.Anything).Return(nil)
+
+	app := app.Application{
+		Services: app.Services{
+			ManageFleet: &service,
+		},
+	}
+
+	handler := NewFleetIDRemovedEventHandler(app)
+
+	requestPb := &skysign_proto.FleetIdRemovedEvent{
+		FleetId: string(DefaultFleetID),
+	}
+	requestBin, _ := proto.Marshal(requestPb)
+	err := handler.HandleFleetIDRemovedEvent(
+		nil,
+		requestBin,
+	)
+
+	a.Nil(err)
+	a.Equal(service.ID, string(DefaultFleetID))
+}
