+++ fleet-formation/pkg/fleet/adapters/rabbitmq/mock_data_test.go
@@ -3,15 +3,51 @@ package rabbitmq
 import (
 	"context"
 	"fleet-formation/pkg/fleet/domain/fleet"
+	"fleet-formation/pkg/fleet/service"
 
 	crm "github.com/Tomofiles/skysign_cloud_v2/skysign-common/pkg/common/adapters/rabbitmq"
 
 	"github.com/stretchr/testify/mock"
 )
 
+const DefaultFleetNumberOfVehicles = 3
 const DefaultFleetID = fleet.ID("fleet-id")
 const DefaultFleetVehicleID = fleet.VehicleID("vehicle-id")
 const DefaultFleetMissionID = fleet.MissionID("mission-id")
+
+type manageFleetServiceMock struct {
+	mock.Mock
+	ID               string
+	NumberOfVehicles int
+	OriginalID       string
+	NewID            string
+}
+
+func (s *manageFleetServiceMock) CreateFleet(
+	command service.CreateFleetCommand,
+) error {
+	ret := s.Called()
+	s.ID = command.GetID()
+	s.NumberOfVehicles = command.GetNumberOfVehicles()
+	return ret.Error(0)
+}
+
+func (s *manageFleetServiceMock) DeleteFleet(
+	command service.DeleteFleetCommand,
+) error {
+	ret := s.Called()
+	s.ID = command.GetID()
+	return ret.Error(0)
+}
+
+func (s *manageFleetServiceMock) CarbonCopyFleet(
+	command service.CarbonCopyFleetCommand,
+) error {
+	ret := s.Called()
+	s.OriginalID = command.GetOriginalID()
+	s.NewID = command.GetNewID()
+	return ret.Error(0)
+}
 
 type channelMockPublish struct {
 	mock.Mock
@@ -44,11 +80,28 @@ func (ch *channelMockPublish) Close() error {
 	panic("implement me")
 }
 
-type publishHandlerMock struct {
+type pubSubManagerMock struct {
+	consumers       []consumer
 	publishHandlers []func(ch crm.Channel, e interface{})
 }
 
-func (h *publishHandlerMock) SetPublishHandler(handler func(ch crm.Channel, e interface{})) error {
+func (h *pubSubManagerMock) SetConsumer(ctx context.Context, exchangeName, queueName string, handler func([]byte)) error {
+	h.consumers = append(
+		h.consumers,
+		consumer{
+			exchangeName: exchangeName,
+			queueName:    queueName,
+			handler:      handler,
+		})
+	return nil
+}
+
+func (h *pubSubManagerMock) SetPublishHandler(handler func(ch crm.Channel, e interface{})) error {
 	h.publishHandlers = append(h.publishHandlers, handler)
 	return nil
+}
+
+type consumer struct {
+	exchangeName, queueName string
+	handler                 func([]byte)
 }
