+++ fleet-formation/pkg/fleet/adapters/rabbitmq/subscribe_event_handler.go
@@ -1,7 +1,8 @@
-package ports
+package rabbitmq
 
 import (
 	"context"
+	"fleet-formation/pkg/fleet/app"
 
 	"github.com/Tomofiles/skysign_cloud_v2/skysign-common/pkg/common/ports"
 
@@ -12,14 +13,17 @@ import (
 func SubscribeEventHandler(
 	ctx context.Context,
 	psm ports.PubSubManagerSetter,
-	evt EventHandler,
+	app app.Application,
 ) {
+	gevt := NewFleetIDGaveEventHandler(app)
+	revt := NewFleetIDRemovedEventHandler(app)
+	cevt := NewFleetCopiedEventHandler(app)
 	psm.SetConsumer(
 		ctx,
 		FleetIDGaveEventExchangeName,
 		FleetIDGaveEventQueueName,
 		func(event []byte) {
-			if err := evt.HandleFleetIDGaveEvent(ctx, event); err != nil {
+			if err := gevt.HandleFleetIDGaveEvent(ctx, event); err != nil {
 				glog.Error(err)
 			}
 		},
@@ -29,7 +33,7 @@ func SubscribeEventHandler(
 		FleetIDRemovedEventExchangeName,
 		FleetIDRemovedEventQueueName,
 		func(event []byte) {
-			if err := evt.HandleFleetIDRemovedEvent(ctx, event); err != nil {
+			if err := revt.HandleFleetIDRemovedEvent(ctx, event); err != nil {
 				glog.Error(err)
 			}
 		},
@@ -39,7 +43,7 @@ func SubscribeEventHandler(
 		FleetCopiedEventExchangeName,
 		FleetCopiedEventQueueName,
 		func(event []byte) {
-			if err := evt.HandleFleetCopiedEvent(ctx, event); err != nil {
+			if err := cevt.HandleFleetCopiedEvent(ctx, event); err != nil {
 				glog.Error(err)
 			}
 		},
