--- fleet-formation/pkg/fleet/ports/event.go
@@ -1,139 +1,0 @@
-package ports
-
-import (
-	"context"
-	"fleet-formation/pkg/fleet/app"
-
-	"github.com/Tomofiles/skysign_cloud_v2/skysign-proto/pkg/skysign_proto"
-
-	"github.com/golang/glog"
-	"google.golang.org/protobuf/proto"
-)
-
-const (
-	// FleetIDGaveEventExchangeName .
-	FleetIDGaveEventExchangeName = "flightplan.fleet_id_gave_event"
-	// FleetIDGaveEventQueueName .
-	FleetIDGaveEventQueueName = "fleet.fleet_id_gave_event"
-	// FleetIDRemovedEventExchangeName .
-	FleetIDRemovedEventExchangeName = "flightplan.fleet_id_removed_event"
-	// FleetIDRemovedEventQueueName .
-	FleetIDRemovedEventQueueName = "fleet.fleet_id_removed_event"
-	// FleetCopiedEventExchangeName .
-	FleetCopiedEventExchangeName = "flightoperation.fleet_copied_event"
-	// FleetCopiedEventQueueName .
-	FleetCopiedEventQueueName = "fleet.fleet_copied_event"
-)
-
-// EventHandler .
-type EventHandler interface {
-	HandleFleetIDGaveEvent(ctx context.Context, event []byte) error
-	HandleFleetIDRemovedEvent(ctx context.Context, event []byte) error
-	HandleFleetCopiedEvent(ctx context.Context, event []byte) error
-}
-
-// eventHandler .
-type eventHandler struct {
-	app app.Application
-}
-
-// NewEventHandler .
-func NewEventHandler(application app.Application) *eventHandler {
-	return &eventHandler{app: application}
-}
-
-// HandleFleetIDGaveEvent .
-func (h *eventHandler) HandleFleetIDGaveEvent(
-	ctx context.Context,
-	event []byte,
-) error {
-	eventPb := skysign_proto.FleetIdGaveEvent{}
-	if err := proto.Unmarshal(event, &eventPb); err != nil {
-		return err
-	}
-
-	glog.Infof("RECEIVE , Event: %s, Message: %s", FleetIDGaveEventQueueName, eventPb.String())
-
-	command := createFleetCommandHolder{
-		event: &eventPb,
-	}
-	if ret := h.app.Services.ManageFleet.CreateFleet(&command); ret != nil {
-		return ret
-	}
-	return nil
-}
-
-// HandleFleetIDRemovedEvent .
-func (h *eventHandler) HandleFleetIDRemovedEvent(
-	ctx context.Context,
-	event []byte,
-) error {
-	eventPb := skysign_proto.FleetIdRemovedEvent{}
-	if err := proto.Unmarshal(event, &eventPb); err != nil {
-		return err
-	}
-
-	glog.Infof("RECEIVE , Event: %s, Message: %s", FleetIDRemovedEventQueueName, eventPb.String())
-
-	command := deleteFleetCommandHolder{
-		event: &eventPb,
-	}
-	if ret := h.app.Services.ManageFleet.DeleteFleet(&command); ret != nil {
-		return ret
-	}
-	return nil
-}
-
-// HandleFleetCopiedEvent .
-func (h *eventHandler) HandleFleetCopiedEvent(
-	ctx context.Context,
-	event []byte,
-) error {
-	eventPb := skysign_proto.FleetCopiedEvent{}
-	if err := proto.Unmarshal(event, &eventPb); err != nil {
-		return err
-	}
-
-	glog.Infof("RECEIVE , Event: %s, Message: %s", FleetCopiedEventQueueName, eventPb.String())
-
-	command := copyCommandHolder{
-		originalID: eventPb.GetOriginalFleetId(),
-		newID:      eventPb.GetNewFleetId(),
-	}
-	if ret := h.app.Services.ManageFleet.CarbonCopyFleet(&command); ret != nil {
-		return ret
-	}
-	return nil
-}
-
-type createFleetCommandHolder struct {
-	event *skysign_proto.FleetIdGaveEvent
-}
-
-func (h *createFleetCommandHolder) GetID() string {
-	return h.event.FleetId
-}
-
-func (h *createFleetCommandHolder) GetNumberOfVehicles() int {
-	return int(h.event.NumberOfVehicles)
-}
-
-type deleteFleetCommandHolder struct {
-	event *skysign_proto.FleetIdRemovedEvent
-}
-
-func (h *deleteFleetCommandHolder) GetID() string {
-	return h.event.FleetId
-}
-
-type copyCommandHolder struct {
-	originalID string
-	newID      string
-}
-
-func (h *copyCommandHolder) GetOriginalID() string {
-	return h.originalID
-}
-func (h *copyCommandHolder) GetNewID() string {
-	return h.newID
-}
