--- fleet-formation/pkg/fleet/ports/event_test.go
@@ -1,105 +1,0 @@
-package ports
-
-import (
-	"fleet-formation/pkg/fleet/app"
-	"testing"
-
-	"github.com/Tomofiles/skysign_cloud_v2/skysign-proto/pkg/skysign_proto"
-
-	"github.com/golang/protobuf/proto"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-)
-
-func TestHandleFleetIDGaveEvent(t *testing.T) {
-	a := assert.New(t)
-
-	service := manageFleetServiceMock{}
-
-	service.On("CreateFleet", mock.Anything).Return(nil)
-
-	app := app.Application{
-		Services: app.Services{
-			ManageFleet: &service,
-		},
-	}
-
-	handler := NewEventHandler(app)
-
-	requestPb := &skysign_proto.FleetIdGaveEvent{
-		FleetId:          DefaultFleetID,
-		NumberOfVehicles: DefaultFleetNumberOfVehicles,
-	}
-	requestBin, _ := proto.Marshal(requestPb)
-	err := handler.HandleFleetIDGaveEvent(
-		nil,
-		requestBin,
-	)
-
-	a.Nil(err)
-	a.Equal(service.ID, DefaultFleetID)
-	a.Equal(service.NumberOfVehicles, DefaultFleetNumberOfVehicles)
-}
-
-func TestHandleDeletedEvent(t *testing.T) {
-	a := assert.New(t)
-
-	service := manageFleetServiceMock{}
-
-	service.On("DeleteFleet", mock.Anything).Return(nil)
-
-	app := app.Application{
-		Services: app.Services{
-			ManageFleet: &service,
-		},
-	}
-
-	handler := NewEventHandler(app)
-
-	requestPb := &skysign_proto.FleetIdRemovedEvent{
-		FleetId: DefaultFleetID,
-	}
-	requestBin, _ := proto.Marshal(requestPb)
-	err := handler.HandleFleetIDRemovedEvent(
-		nil,
-		requestBin,
-	)
-
-	a.Nil(err)
-	a.Equal(service.ID, DefaultFleetID)
-}
-
-func TestHandleFleetCopiedEvent(t *testing.T) {
-	a := assert.New(t)
-
-	var (
-		DefaultFleetOriginalID = DefaultFleetID + "-new"
-		DefaultFleetNewID      = DefaultFleetID + "-new"
-	)
-
-	service := manageFleetServiceMock{}
-
-	service.On("CarbonCopyFleet", mock.Anything).Return(nil)
-
-	app := app.Application{
-		Services: app.Services{
-			ManageFleet: &service,
-		},
-	}
-
-	handler := NewEventHandler(app)
-
-	requestPb := &skysign_proto.FleetCopiedEvent{
-		OriginalFleetId: DefaultFleetOriginalID,
-		NewFleetId:      DefaultFleetNewID,
-	}
-	requestBin, _ := proto.Marshal(requestPb)
-	err := handler.HandleFleetCopiedEvent(
-		nil,
-		requestBin,
-	)
-
-	a.Nil(err)
-	a.Equal(service.OriginalID, DefaultFleetOriginalID)
-	a.Equal(service.NewID, DefaultFleetNewID)
-}
